requirements.txt
aiogram==3.13.1
aiohttp==3.9.5
g4f==0.3.3.2
google_api_python_client==2.149.0
google_auth_oauthlib==1.2.1
protobuf==3.20.2
pydantic==2.9.2
pydantic_settings==2.6.0
Telethon==1.37.0
